apps/router/pyproject.toml
[project]
name = "router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adam McGreggor", email = "adam@adamamyl.com" }
]
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=5.1.0",
    "asyncpg>=0.31.0",
    "emf-shared",
    "fastapi>=0.134.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "bandit[toml]>=1.9.4",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
    "types-python-jose>=3.5.0.20250531",
]

[tool.uv.sources]
emf-shared = { path = "../../shared", editable = true }
